+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # jina-ai-provider
 
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- export JinaEmbeddingOptions
+
+## 1.0.0-beta.0
+
+### Major Changes
+
+- - update to @ai-sdk/provider and @ai-sdk/provider-utils beta
+  - change to zod v4 internally
+
 ## 0.0.2
 
 ### Patch Changes
@@ -11,6 +24,5 @@
 ### Patch Changes
 
 - Initial release of Jina AI Provider with Multimodal Embeddings
-
   - Add support for multimodal embeddings
   - Add support for text embeddings